#F095 ESR - 04.28 - Refatorando a exclusão de cozinhas para usar domain services

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.infrastructure.repository;
 
 import com.algaworks.algafood.domain.model.Cozinha;
 import com.algaworks.algafood.domain.repository.CozinhaRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,6 +13,7 @@ import java.util.List;
 
 @Component
 public class CozinhaRepositoryImpl implements CozinhaRepository {
+
     @PersistenceContext
     private EntityManager entityManager;
 
@@ -23,21 +25,24 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Transactional
     @Override
-    public Cozinha salvar(Cozinha cozinha) {
+    public Cozinha salvar(final Cozinha cozinha) {
         return entityManager.merge(cozinha);
     }
 
     @Transactional
     @Override
-    public Cozinha buscar(Long id) {
+    public Cozinha buscar(final Long id) {
         return entityManager.find(Cozinha.class, id);
     }
 
     @Transactional
     @Override
-    public void remover(Cozinha cozinha) {
+    public void remover(final Long id) {
         // criar uma instancia 'managed' pelo JPA
-        cozinha = buscar(cozinha.getId());
+        Cozinha cozinha = buscar(id);
+        if(cozinha == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(cozinha);
     }
